Fix static linkage of toybox binary.
If toybox was compiled statically e.g. for use in a recovery ramdisk,
it would not build after b66a29a. This is because libselinux already
exports a selinux_log_callback() function which getprop.c is trying
to override.

This change simply makes the second symbol static, since it is only
used in a function table so we do not actually need to export it
from toybox.

Signed-off-by: Alistair Strachan <alistair.strachan@imgtec.com>
Change-Id: I488aab267774f260eff1c7cc071fe40329a07038